createProject: do not create directories and database entry if something important fails

+++ website/ops.py
@@ -1152,27 +1152,25 @@ def getProjectsByUser(user):
 def createProject(project_id, project_name, project_description, project_annotators, project_managers, ref_corpus, src_dict, worflow, langauge, user):
     if projectExists(project_id):
         return {'success': False, "projectID": project_id, "error": "The dict with the entered name already exists"}
-
-    os.makedirs(os.path.join(siteconfig["dataDir"], "projects", project_id))
     if not worflow:
         return {'success': False, "projectID": project_id, "error": "Workflow not specified."}
+    if not os.path.isfile(os.path.join(currdir, 'workflows', worflow, 'Makefile')):
+        return {'success': False, "projectID": project_id, "error": "Can not find workflow template."}
 
+    os.makedirs(os.path.join(siteconfig["dataDir"], "projects", project_id))
     shutil.copytree(os.path.join(currdir, 'workflows', worflow),
                     os.path.join(siteconfig["dataDir"], "projects", project_id), dirs_exist_ok=True)
 
     # get all phases of the project
     all_phases = []
-    try:
-        with open(os.path.join(siteconfig["dataDir"], "projects", project_id, 'Makefile'), 'r') as f:
-            for line in f:
-                res = phases_re.match(line)
-                if res:
-                    all_phases = [x.strip() for x in res.group(1).split(',')]
-    except FileNotFoundError:
-        return {'success': False, "projectID": project_id, "error": "Workflow for project not created correctly."}
+    with open(os.path.join(siteconfig["dataDir"], "projects", project_id, 'Makefile'), 'r') as f:
+        for line in f:
+            res = phases_re.match(line)
+            if res:
+                all_phases = [x.strip() for x in res.group(1).split(',')]
 
-    dicts = {"src_dict": src_dict}
     # for each phase create list of dictionaries
+    dicts = {"src_dict": src_dict}
     for p in all_phases:
         dicts[p] = []
